fix tests

+++ platform_specific/stm32f103/flash_driver.c
@@ -96,7 +96,7 @@ static uint8_t* flashGetPointer() {
     return (uint8_t*) FLASH_START_ADDR;
 }
 
-size_t flashWrite(uint8_t* data, size_t offset, size_t bytes_to_write) {
+int8_t flashWrite(uint8_t* data, size_t offset, size_t bytes_to_write) {
     int8_t status = flashWaitForLastOperation(FLASH_TIMEOUT_VALUE);
 
     if (status < 0) {
